refactor action review projection surface
