+++ control-plane/tests/test_service_persistence_action_reconciliation_review_surfaces.py
@@ -10,6 +10,7 @@ if str(TESTS_ROOT) not in sys.path:
     sys.path.insert(0, str(TESTS_ROOT))
 
 import _service_persistence_support as support
+from aegisops_control_plane import action_review_projection
 from _service_persistence_support import (
     ActionExecutionRecord,
     ActionRequestRecord,
@@ -29,6 +30,62 @@ from _service_persistence_support import (
 )
 
 class ActionReviewSurfacePersistenceTests(ServicePersistenceTestBase):
+    def test_service_delegates_action_review_chain_snapshot_to_projection_module(
+        self,
+    ) -> None:
+        _store, service, promoted_case, evidence_id, reviewed_at = (
+            self._build_phase19_in_scope_case()
+        )
+        observation = service.record_case_observation(
+            case_id=promoted_case.case_id,
+            author_identity="analyst-001",
+            observed_at=reviewed_at,
+            scope_statement="Observed repository permission change requires tracked review.",
+            supporting_evidence_ids=(evidence_id,),
+        )
+        lead = service.record_case_lead(
+            case_id=promoted_case.case_id,
+            observation_id=observation.observation_id,
+            triage_owner="analyst-001",
+            triage_rationale="Privilege-impacting change needs durable business-hours follow-up.",
+        )
+        recommendation = service.record_case_recommendation(
+            case_id=promoted_case.case_id,
+            review_owner="analyst-001",
+            intended_outcome="Review repository owner change evidence before any approval-bound response.",
+            lead_id=lead.lead_id,
+        )
+        action_request = service.create_reviewed_action_request_from_advisory(
+            record_family="recommendation",
+            record_id=recommendation.recommendation_id,
+            requester_identity="analyst-001",
+            recipient_identity="repo-owner-projection-001",
+            message_intent="Notify the accountable repository owner about the reviewed permission change.",
+            escalation_reason="Reviewed GitHub audit evidence requires bounded owner notification.",
+            expires_at=datetime.now(timezone.utc) + timedelta(hours=4),
+            action_request_id="action-request-projection-delegation-001",
+        )
+        record_index = service._build_action_review_record_index()
+        sentinel = {"review_state": "delegated-to-module"}
+
+        with support.mock.patch.object(
+            action_review_projection,
+            "build_action_review_chain_snapshot",
+            autospec=True,
+            return_value=sentinel,
+        ) as build_snapshot:
+            result = service._build_action_review_chain_snapshot(
+                action_request,
+                record_index=record_index,
+            )
+
+        self.assertIs(result, sentinel)
+        build_snapshot.assert_called_once_with(
+            service,
+            action_request,
+            record_index=record_index,
+        )
+
     def test_service_inspects_action_review_states_on_queue_alert_and_case_surfaces(
         self,
     ) -> None:
